+++ lib/mix/tasks/ex2js.dist.ex
@@ -16,7 +16,6 @@ defmodule Mix.Tasks.Ex2js.Dist do
 
     build_standard_library
     copy_artifacts
-    npm_install
     build_tarball
 
     File.rm_rf(@folder_name)
@@ -29,13 +28,7 @@ defmodule Mix.Tasks.Ex2js.Dist do
   defp copy_artifacts() do
     File.mkdir_p(@folder_name <> "/bin")
     File.cp!("ex2js", "#{@folder_name}/bin/ex2js")
-    File.cp!("priv/javascript/code_generator.js", "#{@folder_name}/code_generator.js")
     File.cp_r!("priv/javascript/dist", "#{@folder_name}/lib")
-    File.cp!("priv/javascript/release.package.json", "#{@folder_name}/package.json")
-  end
-
-  defp npm_install() do
-    System.cmd("npm", ["install"], [cd: @folder_name])   
   end
 
   defp build_tarball() do